Allow customisation of SWC config
