+++ package/rules/swc.js
@@ -1,17 +1,13 @@
 const { resolve } = require('path')
 const { realpathSync } = require('fs')
 const { loaderMatches } = require('../utils/helpers')
+const { getSwcLoaderConfig } = require('../swc')
 
 const {
   source_path: sourcePath,
   additional_paths: additionalPaths,
   webpack_loader: webpackLoader
 } = require('../config')
-const { isDevelopment, runningWebpackDevServer } = require('../env')
-
-const isJsxFile = (filename) => !!filename.match(/\.(jsx|tsx)?(\.erb)?$/)
-
-const isTypescriptFile = (filename) => !!filename.match(/\.(ts|tsx)?(\.erb)?$/)
 
 module.exports = loaderMatches(webpackLoader, 'swc', () => ({
   test: /\.(ts|tsx|js|jsx|mjs|coffee)?(\.erb)?$/,
@@ -23,30 +19,5 @@ module.exports = loaderMatches(webpackLoader, 'swc', () => ({
     }
   }),
   exclude: /node_modules/,
-  use: ({ resource }) => ({
-      loader: require.resolve('swc-loader'),
-      options: {
-        jsc: {
-          parser: {
-            dynamicImport: true,
-            syntax: isTypescriptFile(resource) ? 'typescript' : 'ecmascript',
-            [isTypescriptFile(resource) ? 'tsx' : 'jsx']: isJsxFile(resource)
-          },
-          transform: {
-            react: {
-              runtime: 'automatic',
-              development: isDevelopment,
-              refresh: runningWebpackDevServer && isDevelopment
-            }
-          }
-        },
-        sourceMaps: true,
-        env: {
-          coreJs: '3.8',
-          loose: true,
-          exclude: ['transform-typeof-symbol'],
-          mode: 'entry'
-        }
-      }
-    })
+  use: ({ resource }) => getSwcLoaderConfig(resource)
 }))
