+++ package/swc/index.js
@@ -1,0 +1,57 @@
+/* eslint global-require: 0 */
+/* eslint import/no-dynamic-require: 0 */
+
+const { resolve } = require('path')
+const { existsSync } = require('fs')
+const { merge } = require('webpack-merge')
+const { isDevelopment } = require('../env')
+
+const isJsxFile = (filename) => !!filename.match(/\.(jsx|tsx)?(\.erb)?$/)
+
+const isTypescriptFile = (filename) => !!filename.match(/\.(ts|tsx)?(\.erb)?$/)
+
+const getCustomConfig = () => {
+  const path = resolve('config', 'swc.config.js')
+  if (existsSync(path)) {
+    return require(path)
+  }
+  return {}
+}
+
+const getSwcLoaderConfig = (filenameToProcess) => {
+  const customConfig = getCustomConfig()
+  const defaultConfig = {
+    loader: require.resolve('swc-loader'),
+    options: {
+      jsc: {
+        parser: {
+          dynamicImport: true,
+          syntax: isTypescriptFile(filenameToProcess)
+            ? 'typescript'
+            : 'ecmascript',
+          [isTypescriptFile(filenameToProcess) ? 'tsx' : 'jsx']:
+            isJsxFile(filenameToProcess)
+        },
+        transform: {
+          react: {
+            runtime: 'automatic',
+            development: isDevelopment
+          }
+        }
+      },
+      sourceMaps: true,
+      env: {
+        coreJs: '3.8',
+        loose: true,
+        exclude: ['transform-typeof-symbol'],
+        mode: 'entry'
+      }
+    }
+  }
+
+  return merge(defaultConfig, customConfig)
+}
+
+module.exports = {
+  getSwcLoaderConfig
+}
